Fix handling of audio files with more than 2^24 frames.

+++ libaudiofile/aiffwrite.c
@@ -320,7 +320,7 @@ static status WriteSSND (AFfilehandle file)
 	else
 		af_fseek(file->fh, aiff->SSND_offset, SEEK_SET);
 
-	chunkSize = _af_format_frame_size(&track->f, false) *
+	chunkSize = (int) _af_format_frame_size(&track->f, false) *
 		track->totalfframes + 8;
 
 	af_fwrite("SSND", 4, 1, file->fh);
